Kholiday MSVC fix
r1081726 fixed a problem with the MSVC preprocessor, but this was applied in a
file generated by Bison/Flex so was overwritten next time the parser was
regenerated.

This version fixes the grammar in the Bison/Flex source files to no longer use
IN but INOP instead.

CCMAIL: romain@kdab.net

Thanks for catching that Romain!

svn path=/trunk/KDE/kdepimlibs/; revision=1082414