Now that the new holiday data files are all in place, we can turn the old file parser off and fully switch to the new parser.
This includes removing the parallel testing between the old and new parser
which makes the build a lot cleaner as we don't have to try link direct to the
private parser classes.  This caused problems with the Windows build when I
first implemented it early in 4.5, so hopefully this doesn't mess that up.

I'm leaving the old paser code in svn for now, but it will eventually be
deleted.

CCMAIL: romain@kdab.net

svn path=/trunk/KDE/kdepimlibs/; revision=1131705