Revert "add Lebanon holidays"
As discussed in the bug, the change is not complete
and having it as it is in the released Frameworks may be
more problematic than not having it.
Specifically, all the religious date should use the
proper rule for dynamic calculation. This applies
to both Christian holidays like Western and Eastern
Easter and all Islamic holidays, which use a
different calendar.

The initial proposal is interesting but it should go
through a few cycles of reviews, addressing the

CCBUG: 414369
This reverts commit 748d0219f3f34c729f21fe670030a3986c774342.