Avoid parsing each holiday file twice in defaultRegionCode()
Even at twice the speed now, this is still relatively slow and would
probably benefit from either a holiday metadata cache or metadata storage
outside of the holiday files. Technically this is even already mostly
there, as part of the file name.

While at it, also mention unexpected performance costs in the docs.