Replace moon phase calculation code
The Berkeley code had an overflow issue, made accurate calculations
hard, and its license contained an advertising clause. Replace it with
code from the public-domain Moontool.

Since the API of this adapted code gives the phase angle directly, this
also allows to improve and simplify the calculation of the moon phase
for a given date.

Also update the test suite to account for the improved accuracy. The
moon phase dates now match tabulated values from
http://astropixels.com/ephemeris/phasescat/phases2001.html .