Report intermediate lunar phases as well
Those are the phases between the the four specific phase positions already
reported here, see the Wikipedia page mentioned in the docs for details.

The motivation for this is to replace KWeatherCore's sunrise/lunar phase
online query, which is unnecessary given we already have the math for
computing all this ourselves locally here.