Fix reporting invalid sun event times
This was wrongly reporting UTC midnight for non-existing sun events,
such as e.g. during polar day/night.

This was due to a misplaced check, introduced by myself in 27b8283a
during the migration to Qt5.

The QML bindings needs some additional checks, as a QDateTime with an
invalid QTime still results in a valid QDateTime.