Add methods to check for polar days/nights
With !48 (which this depends on for passing its unit tests) we don't
report invalid sunrise/sunset events anymore during polar days/nights,
but we still lacked a way to check whether it's actually day or night
in those cases.

This is needed for example in order to select the correct day or night
weather icons in KWeatherCore.