Turn HolidaysRegionsModel into a table model
This is needed for compatibility with the QML TableView, which unlike its
QQC1 predecessor actually assumes a regular QAIM table model as input, not
a list model with a role to column mapping. The role-based interface
remains, we still need that until the consumer code is ported away from
QQC1.