Introduce SunEvents
Currently, SunRiseSet only provides the times. For Night Light, both
date and time are needed in order to compute timer intervals. The date
cannot be added to the time because the closest sunset or sunrise may
be on another utc date.

Given that the api of the SunRiseSet needs to be changed, this change
uses this opportunity to clean up some things. For example, rename the
SunRiseSet to SunEvents because it is no longer just about sunrise and
sunset. And also allow reusing some calculations.

The SunRiseSet was changed so that it uses the SunEvents behind the
scenes. Some test cases in testsunriseset have been adjusted because
the SunEvents handles julian dates differently, e.g. it passes the time
of solar noon to the equation of time.

The main bulk of code is imported from kwin. Both the SunRiseSet and
the SunEvents use the same formulas to compute dates for solar events.